requirements.txt
asgiref==3.6.0
attrs==22.2.0
backports.zoneinfo==0.2.1
blis==0.7.9
catalogue==2.0.8
certifi==2022.12.7
cffi==1.15.1
chardet==5.1.0
charset-normalizer==3.1.0
click==8.1.3
colorama==0.4.6
confection==0.0.4
cryptography==40.0.1
cymem==2.0.7
Django==4.1.7
django-cors-headers==3.14.0
djangorestframework==3.14.0
docx2txt==0.8
en-core-web-sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.5.0/en_core_web_sm-3.5.0-py3-none-any.whl
idna==3.4
importlib-resources==5.12.0
Jinja2==3.1.2
joblib==1.2.0
jsonschema==4.17.3
langcodes==3.3.0
MarkupSafe==2.1.2
murmurhash==1.0.9
nltk==3.8.1
numpy==1.24.2
packaging==23.0
pandas==1.5.3
pathy==0.10.1
pdfminer.six==20221105
pkgutil-resolve-name==1.3.10
preshed==3.0.8
pycparser==2.21
pycryptodome==3.17
pydantic==1.10.7
pyresparser==1.0.6
pyrsistent==0.19.3
python-dateutil==2.8.2
pytz==2023.3
regex==2023.3.23
requests==2.28.2
six==1.16.0
smart-open==6.3.0
sortedcontainers==2.4.0
spacy==3.5.1
spacy-legacy==3.0.12
spacy-loggers==1.0.4
sqlparse==0.4.3
srsly==2.4.6
thinc==8.1.9
tqdm==4.65.0
typer==0.7.0
typing-extensions==4.5.0
tzdata==2023.3
urllib3==1.26.15
wasabi==1.1.1
zipp==3.15.0
